# Back_end_Flask_Server/openapi.yaml
openapi: 3.0.3
info:
  title: Study-Savvy-Server - OpenAPI 3.0
  description: The API document for Study Savvy.
  termsOfService: https://studysavvy.ddns.net/api
  contact:
    email: open891013@gmail.com
  version: 1.0.0
...
servers:
  - url: https://studysavvy.ddns.net
...
paths:
...
  /api/files:
    get:
      tags:
        - Files
      security:
        - BearerAuth: []
      summary: Return a file's information
      description: File's information can be provided, according to file_id
      parameters:
        - name: page
          in: query
          description: The page number to retrieve.
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_pages:
                    type: integer
                    description: The total number of pages.
                  current_page:
                    type: integer
                    description: The current page number.
                  datas:
                    type: array
                    items:
                      type: object
                      properties:
                        file_time:
                          type: string
                          description: The file time build.
                        file_id:
                          type: string
                          description:  The file ID.
                        status:
                          type: boolean
                          description: The file's current status.
                        file_type:
                          type: integer
                          description: The file type.

        '400':
          description: Invalid status value
        '401':
          description: Not authorize.
        '422':
          description: JWT is invalid.
  /api/files/ASR:
    get:
      tags:
        - Files
      security:
        - BearerAuth: []
      summary: Return a ASR file's information.
      description: The ASR file's information can be provided, according to file_id.
      parameters:
        - name: page
          in: query
          description: The page number to retrieve.
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_pages:
                    type: integer
                    description: The total number of pages.
                  current_page:
                    type: integer
                    description: The current page number.
                  datas:
                    type: array
                    items:
                      type: object
                      properties:
                        file_time:
                          type: string
                          description: The file time build.
                        file_id:
                          type: string
                          description: The file ID.
                        status:
                          type: boolean
                          description: The file's current status.
                        file_type:
                          type: integer
                          description:  The file type.
        '400':
          description: Invalid status value
        '401':
          description: Not authorize.
        '422':
          description: JWT is invalid.
  /api/files/OCR:
    get:
      tags:
        - Files
      security:
        - BearerAuth: []
      summary: Return a OCR file's information.
      description: The OCR file's information can be provided, according to file_id.
      parameters:
        - name: page
          in: query
          description: The page number to retrieve.
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_pages:
                    type: integer
                    description: The total number of pages.
                  current_page:
                    type: integer
                    description: The current page number.
                  datas:
                    type: array
                    items:
                      type: object
                      properties:
                        file_time:
                          type: string
                          description: The file time build.
                        file_id:
                          type: string
                          description: The file ID.
                        status:
                          type: boolean
                          description: The file's current status.
                        file_type:
                          type: integer
                          description: The file type.
        '400':
          description: Invalid status value
        '401':
          description: Not authorize.
        '422':
          description: JWT is invalid.
  /api/files/{file_id}:
    get:
      tags:
        - Files
      security:
        - BearerAuth: []
      summary: Return a file's information.
      description: The file's information can be provided, according to file_id.
      parameters:
        - name: file_id
          in: path
          description: file_id that need to be considered for filter
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: string
                    description: The prompt for OCR.
                  content:
                    type: string
                    description: The file content.
                  summarize:
                    type: string
                    description: The article's summary.
                  details:
                    type: array
                    items:
                      type: string
                      description: The messages for OCR file.
        '400':
          description: Invalid status value
        '401':
          description: Not authorize.
        '403':
          description: Forbidden.
        '404':
          description: Not exist.
        '422':
          description: JWT is invalid.
  /api/files/resources/audio/{file_id}:
    get:
      tags:
        - Files
      security:
        - BearerAuth: [ ]
      summary: Return audio.
      description: Return audio.
      parameters:
        - name: file_id
          in: path
          description: file_id that need to be considered for filter
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid status value
        '401':
          description: Not authorize.
        '403':
          description: Forbidden.
        '404':
          description: Not exist.
        '422':
          description: JWT is invalid.
  /api/files/resources/graph/{file_id}:
    get:
      tags:
        - Files
      security:
        - BearerAuth: [ ]
      summary: Return image.
      description: Return image.
      parameters:
        - name: file_id
          in: path
          description: file_id that need to be considered for filter
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid status value
        '401':
          description: Not authorize.
        '403':
          description: Forbidden.
        '404':
          description: Not exist.
        '422':
          description: JWT is invalid.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT